==> Labs/0_ReviewR.Rmd <==
---
title: "R Review"
author: "Probabalistic Network Modeling"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: html_document
---

```{r include=FALSE}
library(tidyverse)
library(tufte)
library(knitr)
# options(
#   htmltools.dir.version = FALSE, # for blogdown
#   show.signif.stars = FALSE,     # for regression output
#   digits = 2
#   )
#knitr::opts_chunk$set(eval = FALSE)
knitr::opts_chunk$set(echo = TRUE)

```

# Housekeeping

## R/RStudio

You all should have R and RStudio installed on your computers by now. If you do not, first install the latest version of R here: <https://cran.rstudio.com> (remember to select the right installer for your operating system). Next, install the latest version of RStudio here: <https://www.rstudio.com/products/rstudio/download/>. Scroll down to the "Installers for Supported Platforms" section and find the right installer for your operating system.

## R Markdown

You are encouraged to use R Markdown to record your lab work. If you do not already know how to use R markdown, refer to the resources section at the end of this document.

# Getting started

This course will assume that you have some familiarity with the R programming language. However, to refresh your memory on some R key topics, this lab will go through some basic aspects of R functionality that will be important for future exercises.

As a first step, recall how to install new R packages using the R console. One very useful package for working with data is the `tidyverse` package. Try installing it with

```{r, eval = F}
install.packages("tidyverse")
```

Two other packages we will encounter in future labs are `rstan` and `rstanarm`. Install these with the command

```{r, eval = F}
install.packages(c("rstan", "rstanarm"))
```

# Reading and writing files

Depending on who you collaborate with, you may receive data in different forms. These might include Excel (.xls), text files (.txt), comma-separated-value files (.csv), or R DAT files (.dat). One common function we use to load data into the environment is the read.table() function. The call might look like `read.table(file = "file name")`. The file name is fully determined by where the data is located on your device, and where your current working directory is.

Sometimes, we may be able to access data from a website. Let's try to import data from Peter Hoff's "A First Course in Bayesian Statistical Methods":

```{r load_data1}
# for the entire help file for read.table(), type ?read.table into your console
data <- read.table(file = url("http://www2.stat.duke.edu/~pdh10/FCBS/Exercises/azdiabetes.dat"), header = TRUE)
head(data)
```


# Objects

R utilizes objects called data structures, which include numeric vectors, matrices, lists, and data frames. If you are ever unsure about a variable's type, you can run the command `str(<variable>)`.

We will frequently work with vectors and matrices. Vectors can be assigned using the `c()` function (the 'c' stands for 'combine'). We can apply element-wise arithmetic to numeric vectors as follows. Make sure that you understand the following operations!

```{r vectors}
# create numeric vector of length 5
num <- c(1,2,3,4,5)

#the following are equivalent:
num + 1
num + c(1,1,1,1,1)

num + c(1,1) # why does this throw an error?

# multiplication
num2 <- num*2
num2*c(0,0,1,1,1)

# power
num2^2
```

Above, we created a numerical vector by listing out the elements 1,2,3,4,5. This works just fine if we only need a few elements. But what if we want all the integers between 1 and 100? We can use the `seq()` function to generate a numerical vector of values from the specified lower and upper bounds, at some regular interval:

```{r seq}
seq1 <- seq(from = 1, to = 100, by = 1) # the 'by' parameters determines the interval
seq1
```

------------------------------------------------------------------------

### Exercise

1.  Generate a sequence of $100$ equispaced real numbers from $0$ to $1$ and store it in a variable called `seq2`.

------------------------------------------------------------------------

Let's say we want to retrieve the i-th element in a vector. This is called indexing, and we do so using bracket notation. For example, if we want to retrieve the tenth element from seq1, then i=10:

```{r}
seq1[10]
```

If we want the first and last elements, we can index the positions as follows:

```{r}
seq1[c(1,100)]
```

If we want every element except the 25th and 50th, we can easily use the syntax:

```{r}
seq1[-c(25,50)]
```

Suppose we don't want the first element in seq1 to be 1, and would rather change it to 0. We can re-assign elements in numeric vectors by indexing the element we wish to re-assign, and specifying the replacement using `<-`:

```{r}
seq1[1] <- 0
seq1
```

There are many helpful functions which operate on numerical vectors. We can easily calculate the mean, standard deviation, variance, length, and much more! Below are just some examples of functions which you may find useful in this course and moving forward in your statistical career:

```{r vector functions}
seq3 <- seq(from = -3, to = 3, by = .5)
mean(seq3)
sd(seq3)
var(seq3)
length(seq3)
abs(seq3)
exp(seq3)
sqrt(seq3)
is.na(sqrt(seq3))
```

------------------------------------------------------------------------

### Exercise

2.  Sort the entries in `seq3` from greatest to least.

------------------------------------------------------------------------

# Matrices

In this course we will often utilize matrices, which are objects where every row/column is itself a numerical vector. To create a matrix, we use the function matrix() which takes in the data elements, number of columns and rows, and arrangement as arguments. We can then perform matrix arithmetic and operations, such as addition, multiplication, transpose, inverse. For example:

```{r matrix_ops}
mat1 <- matrix(data = seq(from = 1,to = 6, by =1), nrow = 3, ncol = 2, byrow = T)
mat2 <- matrix(data = rep(x = 2, times = 6), nrow = 3, ncol = 2)
mat3 <- mat1+mat2

#transpose of matrix: t()
t(mat3)

# matrix multiplication: %*% 
# make sure dimensions agree!
dim(mat1); dim(mat2)
mat1 %*% t(mat3)

# inverse (if non-singular): solve()
mat4 <- matrix(data = c(1,2,3,4), nrow = 2, ncol = 2, byrow = F)
solve(mat4)

# obtain elements of main diagonal: diag()
diag(mat4)

# create a diagonal matrix: diag(x, nrow, ncol). Default is x=1, which creates an identity matrix
diag(4) # creates 4x4 identity matrix
diag(x = 2, nrow = 4) # creates 4x4 diagonal matrix with 2 on diagonal
```

Indexing matrices is similar to indexing vectors, but now that we are working in 2 dimensions, we need to specify the desired row and column. If we want the element in the i-th row and j-th column of mat4, then we use the syntax mat4[i,j]. If we want the entire i-th row, then we can leave the column index blank: mat4[i,] (and similarly for column).

```{r matrix_ind}
mat4[1,1]
mat4[,2]
mat4[,2] <- c(0,0)
mat4
```

Lastly, suppose we have a large matrix and we wish to find the mean of every column. `apply()` is an extremely useful function which allows you to apply a specified function to an object, either row or column-wise. The function returns a vector or array of values:

```{r apply}
# generate large matrix
mat5 <- matrix(seq(1,100,1), nrow = 4, ncol = 25, byrow = T)

# apply(X = object, MARGIN = 1 for rows or 2 for columns, FUN = function of choice)
# find mean of every column
apply(X = mat5, MARGIN = 2, FUN = mean)
```

------------------------------------------------------------------------

### Exercise

3.  Find the variance of each row of `mat5`

------------------------------------------------------------------------

You can define your own functions and pass them into the FUN argument as well. For example, suppose we want to calculate the natural logarithm of the maximum element in a row/column. We can write a function which takes in a vector and returns the range. We then feed that function in as an argument into `apply()`:

```{r apply_fun}
# create a function to calculate log of maximum for arbitrary x, and returns the value stored in ret
log_max <- function(x) {
  ret <- log(max(x))
  return(ret)
}

# find log of maximum for each column 
apply(X = mat5, MARGIN = 2, FUN = log_max)
```

# Data frames

An R data frame is an array, so the columns/variables can be of different types. If we use the `str()` function to determine the type of object the data we imported is, we notice that the 'data' objects is of type data frame, but the eight variables in the data frame are of different types (int, num, Factor). This differs from matrices, where the elements are solely numerical.

```{r df}
str(data)
```

You may be accustomed to extracting a column of data using the dollar sign operation. We can also use tidyverse language to access/manipulate data frames. However, what is returned from the `select()` function is not a numeric vector but another data frame:

```{r}
# Extract blood pressure from data
bp1 <- data$bp
str(bp1)

# Select blood pressure from data
bp2 <- data %>% select(bp)
str(bp2)
```

In this course we will almost exclusively work with numerical data. Therefore, you may not need to foray into data frames. However, learning the tidyverse syntax can be very useful if you continue to work with R after this course.

# Random number generation and distribution functions

Many aspects of Bayesian inference -- and therefore aspects of this course -- involve simulating random variables from well-known families of distributions. These include the discrete Bernoulli, Poisson, and Binomial families, as well as the continuous Gaussian, Gamma, and Beta distributions.

You can obtain samples from these and other distributions using commands like:

```{r}
# Default is such that the first argument specifies the number of random samples you would like
X <- rnorm(10000, mean = 0, sd = sqrt(2))
Y <- rgamma(10000, shape = 1/2, rate = 1/2)
Z <- rpois(10000, lambda = 5)
```

You can also return the numerical values of the quantiles of these distributions (the inverse of the cumulative distribution function (CDF)) and their probability densities at desired values with slight variations to the base name of the distribution:

```{r}
std_norm_qt <- qnorm(0.95) # For what value x will the CDF function of a N(0,1) R.V. return 0.95?
std_norm_cdf <- pnorm(-2) # What is the value of the CDF function of a N(0,1) R.V. at -2?
std_norm_dens <- dnorm(0.5) # What is the value of the PDF function of a N(0,1) R.V. at 0.5?
```

The distributions you will see in this class almost always come from families that are indexed by one or two parameters. For instance, in the example above, the object `X` contains samples from a Gaussian distribution with mean parameter $\mu = 0$ and variance parameter $\sigma^2 = 2$. Be sure to check the function documentation (with `?rnorm` or a similar command) to see exactly how these functions are parametrized. For instance, as we saw above, the `rnorm` function takes the standard deviation $\sigma$ as its second argument, *not* the variance $\sigma^2$. Special care must also be taken to specify the shape or rate parameterization for the Gamma distribution.

------------------------------------------------------------------------

### Exercise

4.  Generate $500$ samples from a Beta distribution with shape parameter $[a, b] = [0.5, 0.5]$ and store the samples in a variable called `W`

------------------------------------------------------------------------

# Plotting

When making plots in R, you have two main options: (1) the base R plotting function `plot` and (2) the package `ggplot2`.

The base R plotting functions are nice for quick, simple visualizations of data. Here are some examples:

```{r}
norm_samples <- rnorm(10000)
#
par(mfrow = c(2, 2)) # Set the number of rows and columns for display panels
#
hist(norm_samples,
     main = "Base R histogram", 
     xlab = "x", ylab = "Count")
#
plot(x = norm_samples, y = pnorm(norm_samples), 
     main = "Base R scatterplot", 
     xlab = "x", ylab = "Phi(x)")
#
boxplot(norm_samples, 
        main = "Base R boxplot", 
        ylab = "x")
#
plot(density(norm_samples),
     main = "Base R density", 
     xlab = "x", ylab = "Density")
```

Plotting with `ggplot2` is generally a little bit easier when working with data in large tables. It has a gallery of built-in themes, and there are many extensions that make producing complicated visualizations relatively straightforward.

```{r}
norm_samples %>%
  data.frame(x = .) %>%
  ggplot2::ggplot() +
  geom_histogram(aes(x = x, y = ..density..), 
                 fill = "#756bb1", colour = "white", 
                 alpha = 0.5, bins = 30) +
  geom_density(aes(x = x), colour = "#756bb1") +
  geom_vline(aes(xintercept = std_norm_qt)) +
  geom_point(x = 0.5, y = std_norm_dens) +
  labs(x = "x", y = "Density", title = "ggplot density / histogram")
```

------------------------------------------------------------------------

### Exercise

5.  Browse online resources (some below), or use code from above to make a few plots of your own.

------------------------------------------------------------------------

# NIMBLE

NIMBLE is a powerful system for fitting statistical models in \texttt{R}. We will 
formally introduce it on the first day of lecture, but in the meantime it would be 
a good idea to download the necessary packages. Try installing it with

```{r, eval = F}
install.packages("nimble")
install.packages("nimbleHMC")
```

If that doesn't work for you, or you want the latest development version, try

```{r, eval = F}
install.packages("devtools")
library(devtools)
install_github("nimble-dev/nimble", ref = "devel", subdir = "packages/nimble")
```

If you want to get a head start, you can explore the NIMBLE user manual 
[here](https://r-nimble.org/manuals/NimbleUserManual.pdf). 

## Basic NIMBLE example
As a brief introduction, let's go through the Lightning Introduction in Chapter 2
of the NIMBLE manual. But first, one more package: since we'll be performing MCMC, 
the package `bayesplot` is very useful for performing MCMC diagnostics. 

If you haven't already installed this package, do so using the code below. 

```{r, eval = FALSE}
install.package("bayesplot")
```

```{r, echo = FALSE, message = FALSE}
library(bayesplot)
library(nimble)
```

For your first Nimble model, suppose that we have data on $n$ water pumps with $x_i, i = 1, \cdots, n$, recording the number of failures for the $i^{th}$ pump recorded during a time duration of length $t_i$. A natural model for this scenario is to define $x_i$ as a Poisson random variable. 

## Bayesian stats refresher
Quick detour on conjugacy: Recall that in Bayesian statistics, we say that a prior distribution $p(\theta)$is \textit{conjugate} for a given likelihood $p(x \mid \theta)$ if the posterior distribution $p(\theta \mid x)$ has the same distribution family as the  prior. If the likelihood function belongs to the exponential family, then a conjugate prior exists.

Since the Gamma is the conjugate prior for the Poisson likelihood, it's a natural choice for the prior on the rate parameter: 
$$
\begin{aligned}
x_i &\sim \text{Poisson}(\lambda_i) \text{ for } i = 1, \cdots, n, \text{ where } \lambda_i = \theta_i t_i\\
\theta_i &\sim \text{Gamma}(\alpha, \beta) \\
\alpha &\sim \text{Exponential}(1) \\
\beta &\sim \text{Gamma}(0.1, 1)
\end{aligned}
$$

Note that the Exponential and Gamma priors on the parameters $\alpha, \beta$ both
satisfy the requirement that $\alpha, \beta > 0$. 

## Creating the Nimble Model
The general workflow for Nimble is as follows: create a model using BUGS code, 
process that model into a Nimble model, compile the Nimble model into C++ and run. 

To create our model, the first step is to create some BUGS code:

```{r}
pumpCode <- nimbleCode({
for (i in 1:N){
  theta[i] ~ dgamma(shape = alpha, rate = beta)
  lambda[i] <- theta[i]*t[i]
  x[i] ~ dpois(lambda[i])
}
  alpha ~ dexp(rate = 1.0)
  beta ~ dgamma(shape = 0.1, rate = 1.0)
})
```

Now we provide the data, constants, and initial values to perform the MCMC. 

```{r}
pumpConsts <- list(N = 10, 
                   t = c(94.3, 15.7, 62.9, 126, 5.24, 31.4, 1.05, 1.05, 2.1, 10.5))
pumpData <- list(x = c(5, 1, 5, 14, 3, 19, 1, 1, 4, 22))
pumpInits <- list(alpha = 1, beta = 1, theta = rep(0.1, pumpConsts$N))
```

Next, we process our BUGS code, constants, data, and initial values into a NIMBLE model. 
We can also configure the MCMC and check conjugacy; for now we're leaving the default 
configuration as is, bt you can pass additional arguments to configureMCMC if desired. 


```{r}
pump <- nimbleModel(code = pumpCode, name = "pump", constants = pumpConsts,
                    data = pumpData, inits = pumpInits)

configureMCMC(pump, print = TRUE)
```

While the model-specified above incorporates many conjugate relationships (e.g., the Gamma-Gamma is conjugate),
the Gamma(likelihood)-Exponential(prior) is not conjugate. Note that we have a conjugate sampler as expected for $\beta, \theta$, and the default sampler for the non-conjugate $\alpha$ is a randomn walk Metropolis sampler. 

We can view the associated graphical model, showing the dependency structure as follows: 

```{r}
pump$plotGraph()
```

## Running the MCMC
Next, we use Nimble's Markove chain Monte Carlo algorithms to sample from the posterior distribution
defined by our model and the data. The fastest way to fit our model is using the one-line MCMC invocation `nimbleMCMC`.
```{r}
niter <- 5000
nchains <- 2
mcmc.out <- nimbleMCMC(code = pumpCode, constants = pumpConsts,
                       data = pumpData, inits = pumpInits,
                       nchains = nchains, niter = niter,
                       summary = TRUE, WAIC = TRUE,
                       monitors = c('alpha', 'beta', 'theta'))
```

## Posterior Summaries
A nice way to summarize the posterior samples and credible intervals is with the 
`mcmc_intervals` function: 

```{r}
samples <- mcmc.out$samples 
mcmc_intervals(samples)
```

While the above plot gives a nice overview of our posterior samples, we still need 
to determine if our posterior samples are an acceptable approximation of the target
distribution, i.e., we need to evaluate convergence of the MCMC to the stationary
distribution. 

## MCMC Diagnostics
To get some insight into that, we now look at some quick MCMC diagnostics. 
For a quick primer on visual MCMC diagnostics with the package `bayesplot`, visit
the tutorial here: https://mc-stan.org/bayesplot/articles/visual-mcmc-diagnostics.html.


One very useful diagnostic plot is the trace plot, 
a time series plot of the posterior samples from our Markov Chains, this allows us to evaluate the 
evolution of the parameter over time (i.e., "mixing"). In general, a good trace plot should show
random scatter around a mean; if in contrast we see snaking or local trending, it suggests that
our chains have not mixed well and may not yet have converged to the stationary distribution. 
Mixing looks satisfactory in the plots below. 

```{r}
mcmc_trace(samples)
```
The function `mcmc_pairs` allows us to look at multiple parameters at once, and is
useful for identifying collinearity between variables. Note we use only a $\theta_1$
in the plot below for readability; `mcmc_parcoord` works better with a high number 
of parameters. 

```{r}
mcmc_pairs(samples, pars = c("alpha", "beta", "theta[1]"), 
            off_diag_args = list(size = 0.75))
```

The narrow bivariate scatter for $(\alpha, \beta)$ indicates some collinearity. Can you explain this
by looking at the relationship between the parameters of the $\Gamma$ distribution? Start by looking 
at the mean. 

ACF plots show the autocorrelation between successive iterations of a given parameter. 
In the ideal case, our MCMC samples would be independent draws from the posterior, and
so we would like to see ACF dropping quickly to 0 as the lag increases. 

Note that our $\alpha$, sampled via Random Walk has the most persistent autocorrelation, 
and is likely inducing autocorrelation in $\beta$ as well. 



```{r}
mcmc_acf(samples, pars = c("alpha", "beta", "theta[1]"))
```

*** 

### Exercise
Rerun the MCMC above, experimenting with different priors and samplers. 

***

# Saving your output

We can save plots in interactive mode either by saving directly from the plot zoom
pane, or using the `png` command. 

```{r, eval = FALSE}
png(file = "test_save.png")
mcmc_acf(samples, pars = c("alpha", "beta", "theta[1]"))
dev.off() 

```

You can also save the whole fitted model using the  `saveRDS` command, 
and later read it using the `readRDS` command. 

```{r, eval = FALSE}
saveRDS(mcmc.out, file = "test_save.rds")
new_mcmc <- readRDS(mcmc.out)

```

Alternatively, you can simply render this Rmarkdown document using the `Knit`
button in Rstudio to produce a html document or PDF document. 

# R tutorials and resources

For more information on the R programming language please refer to some or all of the following resources:

-   [R for Data Science](https://r4ds.had.co.nz)

-   [R Markdown](https://rmarkdown.rstudio.com/lesson-1.html)

-   [ggplot2](https://ggplot2.tidyverse.org)

-   [ColorBrewer](http://colorbrewer2.org)

-   [Nimble](https://r-nimble.org/manuals/NimbleUserManual.pdf)

-   [Bayesplot](https://mc-stan.org/bayesplot/articles/visual-mcmc-diagnostics.html)

# Acknowledgement

This lab builds on an introductory exercise created by [Jordan Bryan](https://stat.duke.edu/people/jordan-bryan) and [Becky Tang](https://stat.duke.edu/people/becky-tang) as well as the first two chapters of the NIMBLE user manual.
